pyproject.toml
[project]
name = "base-flet"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flet>=0.28.3",
    "httpx>=0.28.1",
    "psycopg2>=2.9.11",
    "pyinstaller>=6.16.0",
    "python-dotenv>=1.2.1",
    "pywin32>=311",
    "pywin32>=308; sys_platform == 'win32'",
    "reportlab>=4.4.4",
    "sqlmodel>=0.0.27",
]
